+++ source/projects/py/CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## [0.2.x]
 
+- Added tests and tested methods for `api.Binbuf` object
+
 - Added [buffer protocol support](https://cython.readthedocs.io/en/latest/src/userguide/buffer.html#buffer) to `api.Buffer` with numpy test
 
 - Combined buffer tests (except numpy and array tests) into `test_buffer.py` and `test_buffer.maxpat`